Extend ceph sc tests
We'll extend the Ceph storage class test to actually create a PVC
and ensure that pods can read/write data from these volumes.

While at it, we're making some slight changes to the ceph bundle.
At the moment, it deploys a single unit with two OSDs and node based
replication (default). For this reason, the PGs are inactive and
the rbd commands hang (and implicitly the Ceph provisioner).
We'll fix this by using three units, each containing one OSD.