Dont bootstrap with containerd_base (#272)
When you bootstrap with a containerd_base_path, this is not the actual path used for containerd.

the k8s snap appends this path to another path[1], so the charms are still not writing configs where you'd expect them to.

If we don't bootstrap with it -- we'll get the correct path by "luck" rather than being specific. Until the snap bug is worked out, this is a workaround

1: https://github.com/canonical/k8s-snap/blob/ea5f8853818984a7a0d766155b744e846d704653/src/k8s/pkg/k8sd/api/cluster_bootstrap.go#L45-L48