feat(config): Prevent changing bootstrap charm config (#319)
### Overview
This PR adds a way to prevent the `bootstrap-` charm configs from getting changed after the cluster is bootstrapped.

### Rationale
Some of the charm config options start with `bootstrap-` which means that these options should only be specified upon deploying the charm (before the underlying Kubernetes cluster is bootstrapped) and changing them afterwards might result in unexpected behavior. We've discussed that we like to have a mechanism to prevent the user from changing these options.